backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.3.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.2
